failed test

+++ app/test/test_metric_stability.py
@@ -36,7 +36,7 @@ img = cv2.imdecode(nparr, cv2.IMREAD_COLOR)
 img = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
 input_h, input_w = 300, 300
 img = cv2.resize(img, (input_w, input_h), interpolation=cv2.INTER_LINEAR)
-img = img.astype(np.uint8)
+img = img.astype(np.float32)
 img = np.expand_dims(img, axis=0)
 
 detections = model.predict(img)
